Added options and ability to set time in number of seconds
The main update was to allow a developer to specify the countdown in
number of seconds rather than forcing the use of a unix timestamp. The
timestamp depends on the user's browser / computer being accurate;
using number of seconds allows the developer to use server-side code to
countdown to a specific time reliably, regardless of the user's
computer's time or timezone.

In order to allow the HTML to be validated, switched the plugin to use
HTML5 data variables, specifically data-time or data-seconds.

Also added options to add a "warning" class to the countdown at a
certain point, so you could turn the countdown red for the last 60
seconds, for example.

Lastly, updated the demo code.